Corrected machine string parsing

+++ Foundation/NSArray+Shuffle.m
@@ -10,7 +10,7 @@
 {
 	NSMutableArray *result = self.mutableCopy;
 	for (NSInteger i = self.count-1; i >= 1; i--)
-		[result exchangeObjectAtIndex:i withObjectAtIndex:arc4random_uniform(i+1)];
+		[result exchangeObjectAtIndex:i withObjectAtIndex:arc4random_uniform((unsigned int)(i+1))];
 	return result;
 }
